Move eLivingModel into GlobalConstants
