+++ GameServer/GlobalConstants.cs
@@ -17,9 +17,7 @@
  *
  */
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using DOL.GS.PacketHandler;
 using DOL.Database;
 using DOL.Language;
 
@@ -1038,6 +1036,61 @@ namespace DOL.GS
         public const string Realm_Spells = "Realm Spells"; // Resurrection illness, Speed of the realm
     }
 
+	public enum eLivingModel : ushort
+	{
+		None = 0,
+		#region AlbionClassModels
+		BritonMale = 32,
+		BritonFemale = 35,
+		HighlanderMale = 39,
+		HighlanderFemale = 43,
+		SaracenMale = 48,
+		SaracenFemale = 52,
+		AvalonianMale = 61,
+		AvalonianFemale = 65,
+		InconnuMale = 716,
+		InconnuFemale = 724,
+		HalfOgreMale = 1008,
+		HalfOgreFemale = 1020,
+		MinotaurMaleAlb = 1395,
+		#endregion
+		#region MidgardClassModels
+		TrollMale = 137,
+		TrollFemale = 145,
+		NorseMale = 503,
+		NorseFemale = 507,
+		KoboldMale = 169,
+		KoboldFemale = 177,
+		DwarfMale = 185,
+		DwarfFemale = 193,
+		ValkynMale = 773,
+		ValkynFemale = 781,
+		FrostalfMale = 1051,
+		FrostalfFemale = 1063,
+		MinotaurMaleMid = 1407,
+		#endregion
+		#region HiberniaClassModels
+		FirbolgMale = 286,
+		FirbolgFemale = 294,
+		CeltMale = 302,
+		CeltFemale = 310,
+		LurikeenMale = 318,
+		LurikeenFemale = 326,
+		ElfMale = 334,
+		ElfFemale = 342,
+		SharMale = 1075,
+		SharFemale = 1087,
+		SylvanMale = 700,
+		SylvanFemale = 708,
+		MinotaurMaleHib = 1419,
+		#endregion
+		#region Hastener
+		AlbionHastener = 244,
+		MidgardHastener = 22,
+		HiberniaHastener = 1910,
+		#endregion Hastener
+	}
+
 	public static class GlobalConstants
 	{
 		private static readonly Dictionary<GameLiving.eAttackResult, byte> AttackResultByte = new Dictionary<GameLiving.eAttackResult, byte>()
@@ -1944,7 +1997,7 @@ namespace DOL.GS
 				eCharacterClass.MaulerHib	 	// Mauler_Hib = 62,
 			}},
 		};
-		
+
 		/// <summary>
 		/// Race to Gender Constraints
 		/// </summary>
