+++ GameServer/gameobjects/PlayerRace.cs
@@ -256,59 +256,4 @@ namespace DOL.GS.Realm
 		DarknessRising = 5,
 		LabyrinthOfTheMinotaur = 6
 	}
-
-	public enum eLivingModel : ushort
-	{
-		None = 0,
-		#region AlbionClassModels
-		BritonMale = 32,
-		BritonFemale = 35,
-		HighlanderMale = 39,
-		HighlanderFemale = 43,
-		SaracenMale = 48,
-		SaracenFemale = 52,
-		AvalonianMale = 61,
-		AvalonianFemale = 65,
-		InconnuMale = 716,
-		InconnuFemale = 724,
-		HalfOgreMale = 1008,
-		HalfOgreFemale = 1020,
-		MinotaurMaleAlb = 1395,
-		#endregion
-		#region MidgardClassModels
-		TrollMale = 137,
-		TrollFemale = 145,
-		NorseMale = 503,
-		NorseFemale = 507,
-		KoboldMale = 169,
-		KoboldFemale = 177,
-		DwarfMale = 185,
-		DwarfFemale = 193,
-		ValkynMale = 773,
-		ValkynFemale = 781,
-		FrostalfMale = 1051,
-		FrostalfFemale = 1063,
-		MinotaurMaleMid = 1407,
-		#endregion
-		#region HiberniaClassModels
-		FirbolgMale = 286,
-		FirbolgFemale = 294,
-		CeltMale = 302,
-		CeltFemale = 310,
-		LurikeenMale = 318,
-		LurikeenFemale = 326,
-		ElfMale = 334,
-		ElfFemale = 342,
-		SharMale = 1075,
-		SharFemale = 1087,
-		SylvanMale = 700,
-		SylvanFemale = 708,
-		MinotaurMaleHib = 1419,
-		#endregion
-		#region Hastener
-		AlbionHastener = 244,
-		MidgardHastener = 22,
-		HiberniaHastener = 1910,
-		#endregion Hastener
-	}
 }
